+++ CHANGELOG.rst
@@ -225,6 +225,8 @@ Added
   `#868 <https://github.com/openego/eGon-data/issues/868>`_
 * Write simBEV metadata to DB table
   `PR #978 <https://github.com/openego/eGon-data/pull/978>`_
+* Add plotting function for installed capacities
+  `#763 <https://github.com/openego/eGon-data/issues/763>`_  
 
 .. _PR #159: https://github.com/openego/eGon-data/pull/159
 .. _PR #703: https://github.com/openego/eGon-data/pull/703